Fix for incorrect time zone conversion
In the tz database, the "Etc\GMT..." names are coded opposite to what
one might expect. For instance, "Etc\GMT-4" is in Abu Dhabi, and not
Detroit. This update fixes the machinery which creates these
"Etc\GMT..." names so that the negative GMT offset is appended to the
name (e.g., Detroit has an offset of -4 hours, so its code is
"Etc\GMT+4").